move flexbox utilities
