+++ components/00-utilities/flexbox/flexbox.config.json
@@ -1,0 +1,3 @@
+{
+  "status": "demo"
+}
