--- components/css_flexbox_utilities/css_flexbox_utilities.config.json
@@ -1,5 +1,0 @@
-{
-	"name": "css flexbox utilities",
-	"label": "CSS Flexbox Utilities",
-	"status": "demo"
-}
