+++ fractal_react/__components.js
@@ -3,7 +3,6 @@
 module.exports = {
   "button": require("../components/button/button.jsx"),
   "checkbox": require("../components/checkbox/checkbox.jsx"),
-  "css-flexbox-utilities": require("../components/css_flexbox_utilities/css_flexbox_utilities.jsx"),
   "css-spacing-utilities": require("../components/css_spacing_utilities/css_spacing_utilities.jsx"),
   "form-group": require("../components/form_group/form_group.jsx"),
   "grid": require("../components/grid/grid.jsx"),
